+++ saw-core-rocq/src/SAWCoreRocq/SpecialTreatment.hs
@@ -353,7 +353,7 @@ sawCorePreludeSpecialTreatmentMap configuration =
 
   -- Utility functions
   ++
-  [ ("id", mapsTo datatypesModule "id")
+  [ ("id", mapsTo sawDefinitionsModule "id")
   ]
 
   -- Natural numbers
@@ -457,8 +457,6 @@ sawCorePreludeSpecialTreatmentMap configuration =
   , ("tail",          mapsTo vectorsModule "tail")
   , ("head_gen",      mapsTo vectorsModule "head_gen")
   , ("tail_gen",      mapsTo vectorsModule "tail_gen")
-  -- workaround to avoid issue translating partially-applied "ite"
-  , ("pmux",          skip)
   ]
 
   -- Streams